agnos_update導入。

+++ launch_chffrplus.sh
@@ -22,6 +22,7 @@ function agnos_init {
 
   # Check if AGNOS update is required
   if [ $(< /VERSION) != "$AGNOS_VERSION" ]; then
+    echo 1 > $DIR/../agnos_update
     AGNOS_PY="$DIR/system/hardware/tici/agnos.py"
     MANIFEST="$DIR/system/hardware/tici/agnos.json"
     if $AGNOS_PY --verify $MANIFEST; then
@@ -44,7 +45,7 @@ function launch {
   # 2. The FINALIZED consistent file has to exist, indicating there's an update
   #    that completed successfully and synced to disk.
 
-  if [ ! -f $DIR/../force_prebuild ]; then
+  if [ ! -f $DIR/../force_prebuild ] && [ ! -f $DIR/../agnos_update ]; then
     rm ${BASEDIR}/.overlay_init
   fi
 
@@ -91,7 +92,7 @@ function launch {
 
   # start manager
   cd system/manager
-  if [ ! -f $DIR/prebuilt ] && [ -f $DIR/../force_prebuild ]; then
+  if [ -f $DIR/../agnos_update ] || [ ! -f $DIR/prebuilt ] && [ -f $DIR/../force_prebuild ]; then
     ./build.py
   fi
   ./manager.py
